refactor package to sailpoint.rdk
