+++ src/test/java/sailpoint/FlattenMultiValuedAttributeTest.java
@@ -5,7 +5,7 @@ import bsh.Interpreter;
 import org.apache.log4j.LogManager;
 import org.apache.log4j.Logger;
 import org.junit.jupiter.api.Test;
-import sailpoint.ps.utils.RuleXmlUtils;
+import sailpoint.rdk.utils.RuleXmlUtils;
 import sailpoint.server.IdnRuleUtil;
 import sailpoint.tools.GeneralException;
